requirements.txt
ccxt>=4.0.0
pandas>=2.0.0
numpy>=1.24.0
torch>=2.0.0
scikit-learn>=1.3.0
xgboost>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
